promptconv - rewiring to work in more environments and act less silly.
I ran it on some of the main culprits for the www2 error messages.  The data's
much cleaner now.

I'm tempted to loop over every file on prod that doesn't pass the "iconv"
test and see if we can't get a clean slate.  Somehow bad data's getting into
our system.


git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@107 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3