Make ^R rerun your diff when you're in the svndiff buffer.
Had to split svndiff into two parts to do this.  It's handy when you
run svndiff and realize that you want to remove this or that change
before checkin, so you use 'gf' to edit those files, and '+' to go
back to the diff.  You can then press ^R to get fresh diff output.

This also required a tweak to vim-pager to set buftype=nofile so it
wouldn't complain if you modified the contents of the pager buffer.  I
figured that was really what we wanted anyway.

git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@134 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3