prv was not working for the case where no args were provided
the list provided to prove was being treated as a single
argument. I'm guessing the (f) things in the creation of
$tests is not doing the desired thing but I couldn't figure
the right thign so faked things out with an echo on the
command line.

also, the regular expression excluding live tests was not
working. find -regex matches the entire path so you need
extra .* stuff




git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@168 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3