At the risk of inciting a holy war, set hard tab width to the one true way for unix.
Ran into this while creating a Makefile.

We don't use hard tabs in any source files, so I don't think this will
be an issue when editing our code.

All of the Perl core sources, and most CPAN authors who still use mixed
tabs and spaces, assume an indent width of 8.

One offender of hard tabs is Damian. He uses hard tabs exclusively,
and we can only assume that he sets his tab width to 4. But then again
we all know his head will be first against the wall when the
revolution happens!

Spaces are the way to go for indenting. When there are hard tabs, they
should be 8.

Anything else... oh yeah... in Chinese 8 is lucky and 4 is death!


git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@179 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3