Add TT2 syntax highlighting, including the cool TT2+HTML mode, and add filetype detection rules that make sense for how we use TT2:
If '[%' appears in the first 3 lines of a *.html file, it is TT2+HTML.

If '[%' appears in the first 3 lines of any other file, it is TT2.

If the higlhighting appears not to catch a file you think it should, it may be
because '[%' doesn't appear until later in the file.  In such a case you
should probably up the number of lines examined and commit.

You may have to rerun stconfig-setup to get this.

This is taken from

  http://www.vim.org/scripts/script.php?script_id=830

git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@224 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3