Add some scripts for creating skin packages.
dir2deb is a generic utility which turns a directory in a .deb file.  I hope
to make a lot more use of this tool, so I put a good bit of effort into making
it do some things I knew I would eventually need.  I actually wrote this for a
WikWed a few months ago, but added some stuff for this skins project.

reorganize-skin-files-for-install.sh is a script which does what you'd think.
Since Subversion won't have the files in the same structure that they get
installed something needs to transform them to that order, and this script
does that.  E.g. in SVN we'd have "wired/{css,javascript,images}" but we
install things as "{css,javascript,images}/wired".  So this script does the
transformation.  It's used by dir2deb (--transform-script) when building the
package.

package-skin.new.  This is the new package-skin script, which will replace the
old one.  I'm not deleting the old one until I've tested more.  It can build a
skin directly from Subversion or, given a directory, can build one like that
too.

I'm not sure stconfig is the right place for these things, but I guess it's
better than nothing.



git-svn-id: https://repo.socialtext.net:8999/svn/stconfig/trunk@437 c8a2641f-bf1a-0410-9bac-bc2d7d3cadf3